requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
gradio==4.44.0
python-dotenv==1.0.1
pydantic==2.7.4

# ── LLM ───────────────────────────────────────────────────────────────────────
anthropic==0.34.2
openai==1.45.0

# ── Embeddings & Reranking ────────────────────────────────────────────────────
sentence-transformers==3.1.0
torch==2.4.0          # CPU-only recommended for HF Spaces free tier
# torch==2.4.0+cpu    # Use this on HF Spaces

# ── Vector Store ──────────────────────────────────────────────────────────────
chromadb==0.5.5

# ── Sparse Search ─────────────────────────────────────────────────────────────
rank-bm25==0.2.2

# ── Document Parsing ──────────────────────────────────────────────────────────
PyMuPDF==1.24.9          # Primary PDF parser
pdfplumber==0.11.4       # Fallback PDF parser
python-docx==1.1.2       # DOCX parser

# ── Utilities ─────────────────────────────────────────────────────────────────
numpy==1.26.4
